Fix Issue #15.
One problem I became aware of, is that while you in principle may just ignore
what characters are unreserved from RFC3986 and use percent-encoding for all
characters, the reality is different.

Many clients disallow the use of a percent-encoded character if there is an
unreserved character with the same value. Hence, percent-encodings and
unreserved characters cannot be used interchangeably.

Second, there was a bug where the padding of small number was wrong. Numbers
like 10 was encoded as %A and not as %0A. It should have been the latter.

This patch fixes the encoded problem encountered.