Fix Issue #17.
Link http_gzip into the tracker delegate. We explicitly tell we support gzip
compression and then handle the message coming back. We then guard against
rest of the system and behave as if the message was received without
compression.

This fixes the issue, since any message coming in with gzip will be handled
correctly.