Make recieving messages "we did not request" more forgiving.
Peers A and B communicate. If A chokes B, then A will delete all
REQUEST messages that B had queued up in A. He will also toss away any
new incoming REQUEST messages.

The interesting thing is if A chokes B, A unchokes B and B sends a REQUEST
message. In this case, whether or not the REQUEST from B gets through depends
on the state of A at the time the REQUEST is recieved. Normally this is not
a problem due to the fibrillation rule: You only alter choke state once every
10 seconds.

However: etorrent is currently highly congesting on the network. If you start
up many torrents and do not have adequate bandwidth, the links will soon
congest heavily. If the OS send queue consist of a PIECE message around 16K
of bytes, and is followed by a REQUEST then the PIECE message acts like a
"cork" on the stream. The REQUEST may then get delayed and the above scenario
is a reality.

We fix this by:

 * Allow a PIECE message to be retrieved more than once. Only decrease the
   amount of missing PIECE messages if it was really new.
 * If we don't know about a PIECE message in our request sets and queues, be
   forgiving.
 * When we get a CHOKE message, we queued up pieces REQUESTed at the remote
   peer to the *front* of the request queue. Now we put them in the back in
   the hope that a "stray" PIECE message become rarer.