Supervise the FS process.
First, rather than adding the filesystem process on the control process, add
it to the supervisor for a single torrent, etorrent_t_sup. Next, rather than
pass the FileDict to the FS process, we can as well add it to the global ETS
table from the control process itself. It doesn't matter if the control process
or the FS process "owns" the ETS data. If any of them dies, the other does
as well, so it doesn't really matter.

This change also opens up for the possibility of shoving more processes into
the etorrent_t_sup, which is pretty good.