Restructure Some of the state code around the ETS map from etorrent_fs_mapper.
First, it simplifies the code quite a bit since data concerning what
pieces we have and do not have is now centered around the code of the
etorrent_fs_mapper. Further, we don't have to pass the rather big "DiskState"
when the State Pid starts up.

We are slowly paving the way for a supervised State process.