Fix Path-metainfo reader code.
I remember I wondered why pathnames were in a list. Now I understand: Each
path component is an item in the list, simply due to the fact that we don't
know how to create path components on different operating systems. Hence,
it is up to the client to assemble path components itself.

Up until now, I always expected the path to be of depth 1, ie a file. That
worked well, also for multiple files until I began hitting directory trees.
Using filename:join, we get that fixed in a portable way as well.