Make code work correctly in supervision trees.
I used exit(shutdown) in a lot of places. This is wrong in so many ways. It
manifests itself by propagating the shutdown message into the supervision tree
and then the whole application shutdowns. Now we are using mutually sent
stop/1 messages in the peer_send/recv modules. It seems like it has stabilized
the system by quite a lot.