Handle stop of torrents gracefully.
When trying to push OTP, do read proc_lib, supervisor and gen_server carefully.
The important thing is that it is only $ancestors that can safely start and
stop other processes in a supervision tree. My puny attempts at using exit/2
were futile.

First, we introduce add_torrent/2 and stop_torrent/2 at etorrent_t_pool_sup.
We identify torrents by the file names in the directory tree in the supervisor,
and can hence stop them with a filename as well.

Then we get rid of crappy attempts at this in etorrent_t_sup.

At last, we alter etorrent_t_manager to use the pools interface for this.

Consequence: Beauty!