Plug a race condition when storing pieces.
If two or more peers wanted to write the same piece, A race-condition
would occur because much of the writing happened outside transaction
context.

The fix here is simple, but paves the way for some more code later
along the same track: serialize writing pieces in the etorrent_fs
process.