Introduce changes to handle *completely insane* trackers.
Here is the complete list so far:
- Need to handle gzip compression. Even if we say we don't support it, many
  trackers doesn't care and happily gzip the data back.
- Need to override the Accept header: Some trackers only allow text/plain or */*
- Need to be case sensitive. The HTTP standard says otherwise, but do the tracker
  implementors care? NO!

Steal the URI parser from the inets application and rename it to avoid clashes.
It is undocumented and may hence change over time, so we'll freeze it in our
repo and then update it as we see fit or when an official URI-parser goes in.

While here, report ourselves as being 0.9dev.