Implement a new choking/unchoking algorithm.
Apart from being much simpler than the old algorithm, it is also more correct.
We utilize ideas loosely taken from BitTornado/Mainline 3.x and Transmission
(avoiding some of the grieveous errors from Transmission). This is far from a
good algorithm, but it is a start.

Before we can go for a better algorithm, we need some more fine-grained rate
measurement than we have now. But then it should be possible to support something
that Mainline 5.2.x, Azureus/Vuze or rtorrent has. This will bring us on par
with their performance.