Restructure connection setup code.
The first nail in the coffin of improving connection setup.

Rather than letting the receiving peer process handle connection setup,
we outsource that task to a spawned process which lives while the
connection is being set up. When it is successfully handshaked, we pass
on TCP socket control to the supervisor tree and then we run.

The code here is rather untested and quite new, but I think it works
somewhat like I would want it to. There are more updates to come along
the same vein, but this is a neat start.