Clean up control process.
The control process relied on raw data, rather than tagged data, in
several places. This kills some of the problems by directly enforcing we
return {value, X} rather than just X.

This provides a more effective hook for the dialyzer and makes it easier
to extend the messaging protocol later on if needed.