Fix dirwatcher sweeping bug.
When we swept over the files, we took the saw to the branch we were
sitting on. If the file were unmarked, we deleted the Key. Now, with the
Key gone, there is no way to find the next Key by a call to ets:next/1.
The fix is to store what the next key is, before operating on the key.
Since there is only one permutation cycle in the iterator-walk, this is
safe.