Add adaptive socket handling.
This change measures the rate of sockets. When the socket is going
beyond 16k / sec in transfer speed incoming, it changes to use the
underlying C code by making the socket active. When it falls below 4k /
sec again, it reverts back to "slow" mode.

There is a nasty complication in the code. Namely that we need to sync
the sender and receiver process on a fast -> slow or slow -> fast mode
change. Note that the sender is the only process which are communicated
with outside the sender and receiver. Hence, it is the sender who is
driving the sync:

1. The receiver notices a transition is needed, fast -> slow or slow ->
fast.

2. The receiver casts to the sender.

3. The sender calls the receiver. Now we have a sync'ed lock.

4. The receiver alters the socket.

5. Both processes enter the state of the new mode.