Prune PieceSets from peers. Fix bug in peer_have/2.
Whenever we get a piece in from a peer, we now check it against our own.
Only if the piece is interesting to us, do we add it to the set of
pieces to consider. Otherwise, we simply skip over it, as it cannot be
interesting.

The same applies with new bitfield messages. Whenever a bitfield message
comes in, we prune its result for all the pieces we already have. Again,
it lets us skip pieces which are not interesting right away when trying
to find chunks to download.