Fix bug in bitfield decoding.
The dialyzer warned about a contract breach due to the new specs added
to the wire proto decoder. It turns out that {value, integer()} was
passed rather than integer() and this is an error. The error was silent
since a less-than (<) comparison were used and that goes well in Erlang.

A static type system would have caught this one right away, but the
dialyzer provides a nice alternative.