Implement code for a new rate management handling.
We now handle the current rate via a simpler algorithm. The send and
receive process at each peer is responsible for maintaining the current
rate. Every 5 seconds, they report this rate (which is a running
average) to a central registry (in ETS). The receiver also keeps track
of a count since the last piece message came by. When more than 30
seconds pass with no PIECE message, we now track and regard that peer as
a "snubbed" peer. This is utilized by the choke manager.

The global rate has been altered. We don't track it anymore. Rather, we
run the global rate by calculation when it is needed. The calculation
is a simple fold over the send and receive tables.