Resolve issue #31: When a torrent completes, tell the tracker.
For some reason, this snuck under the radar while I was otherwise
occupied. Some private trackers finds it rather important that they
get completion messages. While the tracker protocol is *miserable* for
measuring up and download, we can and should report the right kind of
information to the tracker.

All else would be wrong.