Add code for handling BEP 15.
This patch adds a number of new processes to etorrent. These processes
are responsible for handling requests to and from the UDP tracker, when
we later enable it. The code is not yet in a functioning state where it
is linked in, but we need support for the BEP 12/15 coexistence before
we can do this.

Hence, the code is added "as is", with bugs in it. These are to be
corrected later on. While here, introduce the use of EQCmini and eunit
in the code base to begin hunting for bugs.