Implement the chunk_mgr ETS table as a bag rather than a set.
We really wanted this to be a bag instead of a set in semantics. The
change presented here greatly simplifies the code as it changes it from
using sets to bags. We used the set ETS table as a bag anyway, so rather
than handle all the code in Erlang, we can now outsource it to OTP and
the C code.

I expect the change to be much simpler, more readable and easier to
maintain. It will probably also shift the overhead to another kind of
area as per A's law.