Cut number of ets:lookup/2 calls in half in chunk_mgr.
When we search for an applicable piece, we look it up. Then later,
when we assign it, we look up the same piece again. Since we already
hold to reference to the piece in question, just pass the reference on
rather than looking up the value again.

The net effect is simpler code and the number of lookups cut in half.