Document and refactor in the choker code.
We rename some record names and update some variable names. This makes
the code much easier to follow than before. Also while here, begin
writing documentation for each part of the code as we go through
it. The goal is to make the code more comprehensible than it is right
now.