Change the bcoding to be more erlangy.
The tagged code is now more simple: integer() | binary() | [bcode()] |
[{binary(), bcode()}]. There are some problems with this representation
though. First, [] is both the empty list and dictionary. In our world,
that is no problem, but in others worlds it might be so. Hence, we
introduce special syntax, empty_list, empty_dict as atoms in encodings
to remedy this. We explicitly disallow [] as a constructor as it is bad.

While here, carry out eunit/eqc tests all over the place.