Add transmission leeching to the test cases.
We introduce a new test: Seeding from etorrent to a leeching
transmission client. Since we now have more than one test, we shuffle
tests randomly to catch eventual trouble.

Transmission will ignore peers that start with 127.0.0.0/8 but
etorrent will not. So the order is to start etorrent *last* so it
picks up the IP address on the lo interface for transmission and then
connects to it.