Fix the ACK-is-too-old Bug in the code.
It turns out to be two off-by-one bugs were present and made the code
wrongly report that the ACK was too old. The 2nd one, which is the
interesting one, is that the Sequence No. to use is the *next
expected* sequence number. Hence, we should subtract one from it
before we use it to calculate.