Preliminary implementation of false start.
False start is when we begin receiving data before we have a full
connection. This can happen, mostly due to reordering of packets. What
we do is to gather up packets that come in but are not
connection-establishing. These packets then go into a buffer and waits
until we are connected at which point we empty the buffer as a queue.