Fix the RTT Estimation code.
The estimation is in ms, but we are being sent in microseconds. This
of course fails spectacularly. We have a protection on the RTT on 6
seconds. A couple of runs to the moon and back. This guard was tripped
as we were off by a factor of thousand.

The patch fixes this problem.