Remove a bug in the test of piggyback repeated.
The problem was that we forgot to close down sockets after their
use. Hence, the code began to error out since we got duplicate
connection IDs. This happens very quickly, mostly due to the
birthday-paradox.

This patch fixes that and dutifully closes down sockets after they
have been used. I wonder if the reason we sometimes get lockups is due
to things like this.