Enable etorrent_fast_resume:list(), tune sync interval.
We now have a function which allows us to list the contents of the
dets table for fast resume information. While here, tune the sync
interval to 30 seconds. There is no reason to sync every 3 seconds. If
the system crashes, the user can surely accept we may have lost up to
30 seconds of work.