Fix a wrong crypto:rand_uniform/2 call for R14B04.
If the current peer chain was [], we had a call to rand_uniform(0, 0),
which is not allowed. This now fails with a badarg error, as it
should. The patch here fixes this problem by handling the case where
the peer chain is [] specially. In the other case, we had a max call
with 0 as a parameter, but as rand-uniform can return values in the
range Lo =< N > Hi, calling with 0 is ok. Remove this call - also
because there is no explanation of it in the history.