Remove a pattern that will never match (dialyzer)
In the current implementation, the dialyzer has figured out that this
will never match anything. Remove the case for now and quell the
dialyzer so it doesn't report this as an error for now. If it happens,
the code will crash, we will know it, and we can then add back the
case.